refactor: broadly refactor Vault client
This commit broadly refactors the Vault backend to improve K/V lookup for both V1 and V2 KVs. It properly leverages the `prefix` configuration for the Secrets Engine path and accomodates the differences between the KV v1 and KV v2 endpoints, specifically the use of `data` in the secret path. This returns the Vault backend back to pathing parity with the other backends. It also means this commit is a breaking change and is not backwards compatible with configurations that use the more explicit pathing.

This commit also resolves issue #244